+++ libtransmission/torrent-magnet.h
@@ -22,7 +22,7 @@ enum
     METADATA_PIECE_SIZE = (1024 * 16)
 };
 
-void* tr_torrentGetMetadataPiece (tr_torrent * tor, int piece, int * len);
+void* tr_torrentGetMetadataPiece (tr_torrent * tor, int piece, size_t * len);
 
 void tr_torrentSetMetadataPiece (tr_torrent * tor, int piece, const void * data, int len);
