+++ libtransmission/web.h
@@ -68,11 +68,11 @@ struct tr_web_task * tr_webRunWebseed (tr_torrent        * tor,
 
 void tr_webGetTaskInfo (struct tr_web_task * task, tr_web_task_info info, void * dst);
 
-void tr_http_escape (struct evbuffer *out, const char *str, int len, bool escape_slashes);
+void tr_http_escape (struct evbuffer *out, const char *str, size_t len, bool escape_slashes);
 
 void tr_http_escape_sha1 (char * out, const uint8_t * sha1_digest);
 
-char* tr_http_unescape (const char * str, int len);
+char* tr_http_unescape (const char * str, size_t len);
 
 #ifdef __cplusplus
 }
